Remove utils/hooks file and move the hook to src/hooks
